python-api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
langchain-google-genai==0.0.6
supabase==2.3.7
python-dotenv==1.0.0